ai controller updates

+++ Source/TresGame/Public/TresEnemyXIIILPawnBase.h
@@ -15,7 +15,7 @@ public:
     USQEX_ParticleAttachDataAsset* m_UniqueEffectDataAsset;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere, meta=(AllowPrivateAccess=true))
-    uint8 m_bEnableRevengeChain: 1;
+    uint32 m_bEnableRevengeChain: 1;
     
     UPROPERTY(BlueprintReadWrite, EditAnywhere, meta=(AllowPrivateAccess=true))
     TArray<FName> m_RevengeChainBaseDataTableIDNameList;
